% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA.R
\name{CARMA}
\alias{CARMA}
\title{CAusal Robust Mapping Method with Annotations(CARMA)}
\usage{
CARMA(
  z.list,
  ld.list,
  w.list = NULL,
  lambda.list,
  tau = NULL,
  effect.size.prior = "Cauchy",
  rho.index = 0.99,
  BF.index = 10,
  Max.Model.Dim = 20000,
  all.iter = 10,
  all.inner.iter = 10,
  label.list = NULL,
  output.labels = ".",
  input.alpha = 0,
  epsilon.threshold = 0.001,
  input.prior.prob = NULL
)
}
\arguments{
\item{z.list}{Input list of the summary statistics of the testing loci, and each element of the list is the summary statistics of each individual locus.}

\item{ld.list}{Input list of the LD correlation matrix of the testing loci, and each element of the list is the LD matrix of each individual locus.}

\item{w.list}{Input list of the functional annotations of the testing loci, and each element of the list is the functional annotation matrix of each individual locus.}

\item{lambda.list}{Input list of the hyper-parameter \eqn{\eta} of the testing loci, and each element of the list is the hyper-parameter of each individual locus.}

\item{effect.size.prior}{The prior of the effect size. The choice are 'Cauchy', 'Hyper-g', and 'Normal' priors, where the Cauchy prior is the default prior.}

\item{rho.index}{A number between 0 and 1 specifying \eqn{\rho} of the estimated credible sets.}

\item{BF.index}{A number greater than 1 to specifying the threshold of the Bayes factor of the estimated credible models.}

\item{Max.Model.Dim}{Maximum number of the top candidate models based on the unnormalized posterior probability.}

\item{all.iter}{Maximum iterations for EM algorithm to run.}

\item{all.inner.iter}{Maximum iterations for Shotgun algorithm to run per iteration within EM algorithm.}

\item{label.list}{Input list of the names of the testing loci. Default is NULL.}

\item{output.labels}{Output directory where output will be written while CARMA is running. Default is the root directory '.'.}

\item{input.alpha}{The elastic net mixing parameter, where \eqn{0\le}\eqn{\alpha}\eqn{\le 1}.}

\item{epsilon.threshold}{Convergence threshold measured by average of Bayes factors.}
}
\value{
The form of the return is a list, for each list:
\itemize{
\item pip - The posterior inclusion probability of each individual locus.
\item Credible set - The information regarding the credible set given a threshold \eqn{\rho}.
\item Credible model - The information regarding the credible model given a threshold  of the Bayes factor.
}
}
\description{
Performs a Bayesian fine-mapping model in order to identify putative causal variants at GWAS loci. This function requires the summary statistics
of the SNPs in the testing loci, and the corresponding simple-LD matrices for fine-mapping. Functional annotations can be included as the prior 
information of the causality of the testing SNPs. The model can be executed chromosome-wise to increase power.
}
\details{
The function performs a Bayesian fine-mapping method.
}
\examples{
# Example 
set.seed(1)
n = 400
p = 500
beta = rep(0,p)
beta[1] = 1
X = matrix(rnorm(n*p),nrow = n,ncol = p)
X = scale(X,center = TRUE,scale = TRUE)
y = drop(X \%*\% beta + rnorm(n))
SS=compute_summary_statistics(X,y)
z.list<-list()
z.list[[1]]<-(SS$betahat/SS$sebetahat)
ld.list<-list()
ld.list[[1]]<-cov(X)
lambda.list<-list()
lambda.list[[1]]<-1/sqrt(p)
CARMA.result<-CARMA(z.list,ld.list=ld.list,lambda.list = lambda.list,effect.size.prior='Hyper-g')
}
