% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARMA_fixed_sigma_pro.R
\name{CARMA_fixed_sigma_pro}
\alias{CARMA_fixed_sigma_pro}
\title{CARMA (fixed variance)}
\usage{
CARMA_fixed_sigma_pro(
  z.list,
  ld.list,
  w.list = NULL,
  lambda.list = NULL,
  output.labels = NULL,
  label.list = NULL,
  effect.size.prior = "Cauchy",
  rho.index = 0.99,
  BF.index = 10,
  Max.Model.Dim = 10000,
  all.iter = 10,
  all.inner.iter = 10,
  input.alpha = 0.5,
  epsilon.threshold = 0.001,
  num.causal = 10,
  y.var = 1,
  outlier.switch = T,
  outlier.BF.index = 10
)
}
\arguments{
\item{z.list}{Input list of the summary statistics of the testing loci, and each element of the list is the summary statistics of each individual locus.}

\item{ld.list}{Input list of the LD correlation matrix of the testing loci, and each element of the list is the LD matrix of each individual locus.}

\item{w.list}{Input list of the functional annotations of the testing loci, and each element of the list is the functional annotation matrix of each individual locus.}

\item{lambda.list}{Input list of the hyper-parameter \eqn{\eta} of the testing loci, and each element of the list is the hyper-parameter of each individual locus.}

\item{output.labels}{Output directory where output will be written while CARMA is running. Default is NULL.}

\item{label.list}{Input list of the names of the testing loci. Default is NULL.}

\item{effect.size.prior}{The prior of the effect size. The choice are 'Cauchy', 'Hyper-g', and 'Normal' priors, where the Cauchy prior is the default prior.}

\item{rho.index}{A number between 0 and 1 specifying \eqn{\rho} of the estimated credible sets.}

\item{BF.index}{A number greater than 1 to specifying the threshold of the Bayes factor of the estimated credible models.}

\item{Max.Model.Dim}{Maximum number of the top candidate models based on the unnormalized posterior probability.}

\item{all.iter}{Maximum iterations for EM algorithm to run.}

\item{all.inner.iter}{Maximum iterations for Shotgun algorithm to run per iteration within EM algorithm.}

\item{input.alpha}{The elastic net mixing parameter, where \eqn{0\le}\eqn{\alpha}\eqn{\le 1}.}

\item{epsilon.threshold}{Convergence threshold measured by average of Bayes factors.}

\item{num.causal}{The maximum number of causal variants assumed per locus, which is 10 causal SNPs per locus by default.}

\item{y.var}{The input variance of the summary statistics, the default value is 1 as the summary statistics are standardized.}

\item{outlier.switch}{The indicator variable of whether turn on the outlier detection. We suggest that the detection should always turn on if using external LD matrix.}

\item{outlier.threshold}{The Bayes threshold of the hypothesis testing of determining outliers, which is 10 by default.}
}
\value{
The form of the return is a list, for each list:
\itemize{
\item pip - The posterior inclusion probability of each individual locus.
\item Credibleset - The information regarding the credible set given a threshold \eqn{\rho}.
\item Credible model - The information regarding the credible model given a threshold  of the Bayes factor.
\item Outliers - The information regarding the detected outliers.
}
}
\description{
Performs a Bayesian fine-mapping model in order to identify putative causal variants at GWAS loci. The model requires the summary statistics
of the SNPs in the testing loci, the corresponding LD matrices for fine-mapping, and an estimated variance of traits. Functional annotations can be included as the prior 
information of the causality of the testing SNPs. The model also provides a procedure of outlier detection, which resolves the discrepancies
between the summary statistics and the LD matrix extracted from reference panels. The model can be executed chromosome-wise to increase power.
}
\details{
The function performs a Bayesian fine-mapping method.
}
\examples{
# Example 
set.seed(1)
n = 400
p = 500
beta = rep(0,p)
beta[1] = 1
X = matrix(rnorm(n*p),nrow = n,ncol = p)
X = scale(X,center = TRUE,scale = TRUE)
y = drop(X \%*\% beta + rnorm(n))
SS=compute_summary_statistics(X,y)
z.list<-list()
z.list[[1]]<-(SS$betahat/SS$sebetahat)
ld.list<-list()
ld.list[[1]]<-cov(X)
lambda.list<-list()
lambda.list[[1]]<-1/sqrt(p)
CARMA.result<-CARMA_fixed_sigma(z.list,ld.list=ld.list,
lambda.list = lambda.list,effect.size.prior='Hyper-g')
}
