% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hyper_g_fixed_sigma_marginal}
\alias{hyper_g_fixed_sigma_marginal}
\title{Marginal likelihood of hyper-g prior when varinace is fixed}
\description{
Marginal likelihood of hyper-g prior when varinace is fixed
}
