% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Normal_fixed_sigma_marginal}
\alias{Normal_fixed_sigma_marginal}
\title{Marginal likelihood of Normal prior when varinace is fixed}
\description{
Marginal likelihood of Normal prior when varinace is fixed

Marginal likelihood of Normal prior when varinace is fixed

Marginal likelihood of Normal prior when varinace is fixed

Marginal likelihood of Normal prior when varinace is fixed
}
